requirements.txt
pyspark>=3.5,<4.0
delta-spark>=3.2,<4.0
kafka-python>=2.0,<3.0
streamlit>=1.32,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
pyarrow>=15,<20
plotly>=5.19,<6.0
great-expectations>=1.0,<2.0
python-dotenv>=1.0,<2.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
